Separated the Modal for Editing from MyFlashcard.tsx
