+++ src/pages/MyFlashcard/MyFlashcardsComponent.tsx
@@ -1,0 +1,1 @@
+export default function MyFlashcardsComponent() {}
